requirements.txt
Pillow<7
tqdm
numpy
torch
torchvision
opencv-python
tensorflow-gpu
